Shebang portability fix
Add the -S option to `env` to split the parameter on Linux. Without
this, the command is executed as `env "node --experimental-modules"`,
which results in the following error:

```
env: ‘node --experimental-modules’: No such file or directory
env: use -[v]S to pass options in shebang lines
```

The -S option is available on Linux as of GNU coreutils 8.30 [1] (July
2018), and, according to this [2], also works with the MacOS (BSD)
`env`.

[1] https://savannah.gnu.org/forum/forum.php?forum_id=9187
[2] https://apple.stackexchange.com/a/297535